api/src/requirements_grpc.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

pydantic
# Additional utilities
numpy<2.0.0

git+https://github.com/rayliuca/gector.git